Add test describing regression
For fields with multiple values, the `value` cached property
must take non-scalar raw_values into account. Effectively,
`raw_value = (11, 22)` should become `value = ('11', '22')` and
not `value = '(11, 22)'`. When a choice exists with the value
a substring (after force_text) of a selected choice, this first
choice will incorrectly be marked as selected because the string
is contained the string `value`.