For DateInput, DateTimeInput and TimeInput Django uses input_type 'text'. Align (#62)
sniplates with this as well.

Using 'date', 'time' or 'datetime' is problematic because the HTML standard
defines specific formatting for these input types. For example type 'date'
requires a formatting of to be YYYY-MM-DD to be set in the 'value' attribute of
the input element.

However the DateInput, DateTimeInput and TimeInput widgets use
DATE_INPUT_FORMATS, DATETIME_INPUT_FORMATS and TIME_INPUT_FORMATS respectively
to format the value.  And these settings can, and often do, differ from what is
required by browser/HTML 5.